refactor(headless-styles): use common 'Tech' type
