+++ packages/headless-styles/src/components/ErrorMessage/types.ts
@@ -1,10 +1,8 @@
+import type { Tech } from '../types'
+
 export interface ErrorMessageOptions {
   id: string
   invalid: boolean
   message: string
   tech?: Tech
 }
-
-// types
-
-export type Tech = 'svelte'
