+++ packages/headless-styles/src/components/FormLabel/formLabelCSS.ts
@@ -1,5 +1,5 @@
 import { createClassProp } from '../../utils/helpers'
-import { Tech } from '../types'
+import type { Tech } from '../types'
 import { getDefaultFormLabelOptions, getFormValue } from './shared'
 import styles from './formLabelCSS.module.css'
 import { FormLabelOptions } from './types'
